+++ src/gui/controls/progresswidget.h
@@ -36,7 +36,7 @@ class Track;
 
 namespace Player {
 class PlayerManager;
-enum PlayState : uint8_t;
+enum PlayState : int;
 } // namespace Player
 } // namespace Core
